fix: return the insert id as uint64_t so event ids past 2^31 survive
mysql_insert_id() returns my_ulonglong, but zmDbDoInsert() returned int.
Once Events.Id passed INT32_MAX the id was truncated, then sign-extended
when stored in Event::id (uint64_t), so 3502167073 became
18446744072916751373. The Events_Hour/Day/Week/Month inserts were given
the negative value and rejected, and every later UPDATE of the event
matched no row.

Also widen the callers that re-truncated the result: the re-analysis
event id in zma, Tag::Id() when building an Event_Tag, and
Event_Tag::save(). The Events_* insert format specifiers become PRIu64
to match the unsigned id.

fixes #4974

+++ src/zm_db.cpp
@@ -22,6 +22,7 @@
 #include "zm_signal.h"
 #include <algorithm>
 #include <cctype>
+#include <cinttypes>
 #include <cstdlib>
 #include <unistd.h>
 
@@ -267,7 +268,7 @@ int zmDbDo(const std::string &query) {
   return 1;
 }
 
-int zmDbDoInsert(const std::string &query) {
+uint64_t zmDbDoInsert(const std::string &query) {
   std::lock_guard<std::mutex> lck(db_mutex);
   if (!zmDbConnected and !zmDbConnect())
     return 0;
@@ -282,9 +283,8 @@ int zmDbDoInsert(const std::string &query) {
         return 0;
     }
   }
-  // Might not be an int... FIXME
-  int id = mysql_insert_id(&dbconn);
-  Debug(2, "Success running sql insert %s. Resulting id is %d", query.c_str(), id);
+  uint64_t id = mysql_insert_id(&dbconn);
+  Debug(2, "Success running sql insert %s. Resulting id is %" PRIu64, query.c_str(), id);
   return id;
 }
